Generalize office-viewer-url into generic signed-url endpoint
Move the Office viewer URL construction from the server to the client
component, making the endpoint a generic signed-url provider. This
removes project-specific restrictions and content type guards from
the server, relying on space membership for access control and
client-side gating for viewer compatibility.
